+++ src/BracketsEvents.js
@@ -2,12 +2,16 @@ define(function (require) {
     "use strict";
 
     // Brackets modules
-    var FileSystem    = brackets.getModule("filesystem/FileSystem");
+    var DocumentManager = brackets.getModule("document/DocumentManager"),
+        FileSystem      = brackets.getModule("filesystem/FileSystem"),
+        ProjectManager  = brackets.getModule("project/ProjectManager");
 
     // Local modules
     var Events        = require("src/Events"),
         EventEmitter  = require("src/EventEmitter"),
-        Git           = require("src/git/Git");
+        Git           = require("src/git/Git"),
+        HistoryViewer = require("src/HistoryViewer"),
+        Utils         = require("src/Utils");
 
     function refreshStatus() {
         // Extension parts should listen to GIT_STATUS_RESULTS
@@ -25,11 +29,43 @@ define(function (require) {
     EventEmitter.on(Events.GIT_ENABLED, function () {
         attachGitOnlyEvents();
     });
+
     EventEmitter.on(Events.GIT_DISABLED, function () {
         detachGitOnlyEvents();
     });
-    FileSystem.on("change", function (evt, file) {
-        EventEmitter.emit(Events.BRACKETS_FILE_CHANGED, evt, file);
+
+	FileSystem.on("change", function (evt, file) {
+        // we care only for files in current project
+        if (file.fullPath.indexOf(Utils.getProjectRoot()) === 0) {
+            EventEmitter.emit(Events.BRACKETS_FILE_CHANGED, evt, file);
+        }
+    });
+
+    $(DocumentManager).on("documentSaved", function (evt, doc) {
+        // we care only for files in current project
+        if (doc.file.fullPath.indexOf(Utils.getProjectRoot()) === 0) {
+            EventEmitter.emit(Events.BRACKETS_DOCUMENT_SAVED, evt, doc);
+        }
+    });
+
+    $(DocumentManager).on("currentDocumentChange", function (evt, currentDocument, previousDocument) {
+        currentDocument = currentDocument || DocumentManager.getCurrentDocument();
+        if (!HistoryViewer.isVisible()) {
+            EventEmitter.emit(Events.BRACKETS_CURRENT_DOCUMENT_CHANGE, evt, currentDocument, previousDocument);
+        }
+    });
+
+    $(ProjectManager).on("projectOpen", function () {
+        EventEmitter.emit(Events.BRACKETS_PROJECT_CHANGE);
+    });
+
+    $(ProjectManager).on("projectRefresh", function () {
+        EventEmitter.emit(Events.BRACKETS_PROJECT_REFRESH);
+    });
+
+    $(ProjectManager).on("beforeProjectClose", function () {
+        // Disable Git when closing a project so listeners won't fire before new is opened
+        EventEmitter.emit(Events.GIT_DISABLED);
     });
 
 });